chore: extract scale color util
